requirements.txt
# 数据处理（全Python版本兼容，无冲突）
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2

# 环境配置
python-dotenv==1.0.0

# 通义千问SDK（真实存在的最新稳定版）
dashscope==1.25.17

# ========================
# LangChain 1.0+ 官方稳定版（无兼容错误）
# ========================
langchain==1.0.0
langchain-core==0.1.23
langchain-community==0.2.0
langchain-text-splitters==0.1.0

# 向量库（稳定兼容）
faiss-cpu==1.8.0